♻️ refactor(monitoring): extract codex inspection settings
Move Codex inspection setting defaults, normalization, and browser setting persistence into a dedicated model module.

The core inspection service now imports shared readers and settings helpers; validated with type-check, codex inspection tests, and lint.
